# api-contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic AI Agricultural Crop Planning & Precision Farming System",
    "summary": "Agentic multi-agent crop planning and precision farming API.",
    "description": "\nMulti-agent decision support for crop planning and precision farming.\n\nThe workflow is orchestrated with **LangGraph** over twelve specialised agents\n(farm profile, soil & nutrient, weather & climate, agricultural knowledge RAG,\ncrop suitability, ML forecasting, irrigation, crop-risk review, activity\nplanning and advisory). Every agent emits structured evidence, and every\nphysical action (irrigation in particular) requires **human approval** before it\ncan leave the `planned` state.\n",
    "contact": {
      "name": "Precision Farming Platform"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/farms": {
      "get": {
        "tags": [
          "farms & fields"
        ],
        "summary": "List farms",
        "operationId": "list_farms_api_v1_farms_get",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FarmOut"
                  },
                  "title": "Response List Farms Api V1 Farms Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Register a farm",
        "operationId": "create_farm_api_v1_farms_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/farms/{farm_id}": {
      "get": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Get a farm with its fields",
        "operationId": "get_farm_api_v1_farms__farm_id__get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmOutWithFields"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Update a farm",
        "operationId": "update_farm_api_v1_farms__farm_id__patch",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Delete a farm and its fields",
        "operationId": "delete_farm_api_v1_farms__farm_id__delete",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/farms/{farm_id}/summary": {
      "get": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Dashboard summary for a farm",
        "operationId": "farm_summary_api_v1_farms__farm_id__summary_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fields": {
      "get": {
        "tags": [
          "farms & fields"
        ],
        "summary": "List fields",
        "operationId": "list_fields_api_v1_fields_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farm Id"
            }
          },
          {
            "name": "soil_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Soil Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldOut"
                  },
                  "title": "Response List Fields Api V1 Fields Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Add a field to a farm",
        "operationId": "create_field_api_v1_fields_post",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Farm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fields/counts": {
      "get": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Field counts and totals",
        "operationId": "field_counts_api_v1_fields_counts_get",
        "parameters": [
          {
            "name": "farm_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Farm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Field Counts Api V1 Fields Counts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fields/{field_id}": {
      "get": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Get a field",
        "operationId": "get_field_detail_api_v1_fields__field_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Update a field",
        "operationId": "update_field_api_v1_fields__field_id__patch",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "farms & fields"
        ],
        "summary": "Delete a field",
        "operationId": "delete_field_api_v1_fields__field_id__delete",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FarmCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "examples": [
              "Sanjana Rice Farm"
            ]
          },
          "owner_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Owner Name"
          },
          "location_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Location Name",
            "examples": [
              "Kolanupalli, Kakinada"
            ]
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude",
            "examples": [
              16.99
            ]
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude",
            "examples": [
              82.25
            ]
          },
          "village": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Village"
          },
          "district": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "state": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "total_area_ha": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 10000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Area Ha"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "owner_name",
          "location_name"
        ],
        "title": "FarmCreate"
      },
      "FarmOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "location_name": {
            "type": "string",
            "title": "Location Name"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "village": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Village"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "total_area_ha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Area Ha"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "field_count": {
            "type": "integer",
            "title": "Field Count",
            "default": 0
          },
          "total_registered_area_ha": {
            "type": "number",
            "title": "Total Registered Area Ha",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_name",
          "location_name",
          "latitude",
          "longitude",
          "village",
          "district",
          "state",
          "total_area_ha",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "FarmOut"
      },
      "FarmOutWithFields": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "location_name": {
            "type": "string",
            "title": "Location Name"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "village": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Village"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "total_area_ha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Area Ha"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "field_count": {
            "type": "integer",
            "title": "Field Count",
            "default": 0
          },
          "total_registered_area_ha": {
            "type": "number",
            "title": "Total Registered Area Ha",
            "default": 0.0
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldOut"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_name",
          "location_name",
          "latitude",
          "longitude",
          "village",
          "district",
          "state",
          "total_area_ha",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "FarmOutWithFields"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}